requirements.txt
flask
tensorflow
numpy
pillow
matplotlib
